backend/requirements.txt
Flask==3.0.0
Flask-RESTful==0.3.10
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.6.0
Flask-Migrate==4.0.5
Flask-Mail==0.9.1
python-dotenv==1.0.0
PyMySQL==1.1.0
cryptography==41.0.7
pandas==2.1.4
numpy==1.26.2